Update index.html
Restructured the layout: now the options sit on the left and the graph is to the right. This incorporates the 1 : 1.618 (golden) ratio which looks good on the eye.
Added a whole bunch of options to increase interactivity and UX.
Altered the SVG element to use the viewbox attribute instead of defining the height and width; this allows me to scale the box without using clunky, step-by-step media queries and without losing half the display (when changing height and width style properties).