Update machine-learning-chart.js
GRAPH_X/Y are now set to be 600 (simply the width/height I was using before) but the graph's actual size is immediately set by the function setGraphSize. This allows the app to respond well to different screen sizes by setting the width to 100% of the available space and setting the height to equal the graph-box's offsetWidth (to keep it square - SVG's don't have height/width properties or attributes unless defined so offsetWidth had to be used).
I added the interval variable so that the user can change it to watch how the AI learns step-by-step.
Declared a resetVars function to avoid repeating myself whenever the loop is finished.
Also added the renderLastResult function to put the last result into the result-display input box, although I'm not happy with the styling.