Update machine-learning-chart.js
Took some const variables such as RANDOM_WEIGHTS and LEARNING_RATE and made them into a function which returns the same thing as before. This allows me to use the same variables but also to reassign them with the user input, increasing interactivity and overall UX.
Added functions to deal with the new toggle switches and user inputs.
Added an initialiseFunction function so that the first render on the graph is with the purely-random weights rather than the first iteration of trained weights. You can view this by setting the Max Iterations input to 0.
Changed the way the learning rate is written and functions: before, there was a hard-coded chained ternary saying if the iteration count was less than a/b/c then use x/y/z as a learning rate. Instead, there is a 2D array in which each element is written as [iteration-step, learning-rate]. The user can create their own array from the boxes provided and, using a separate learnRateCount incrementer, the function will cycle through the learningRate array.