Update machine-learning-chart.js
Added colours to the variable declaration for much easier styling later on (potentially giving the user access to the colour palette).
Wrote a huge function in which each point is read and then mapped onto the summary-graph. This project was my first introduction to SVG elements and learning about Namespace and how SVG children elements can't have children, that you can't append HTML to SVGs without using foreignObject tags etc was a big task!
Changed the range of functions which update the input values to use one function, setVariable, and to take the desired variable name as an argument instead. At first I thought I would have to use eval or window[varName] to change the variable from a string, but this works just as well.